LPD-47730 Move RelatedModel class to object-api to use it in the test
