+++ modules/apps/object/object-api/src/main/resources/com/liferay/object/entries/frontend/data/set/data/model/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
